+++ src/app/contextual-action-panel/contextual-action-panel-tabs-vertical.component.ts
@@ -1,0 +1,26 @@
+import {
+  Component,
+  HostBinding,
+} from '@angular/core';
+
+// @ts-ignore
+import { SohoContextualActionPanelRef } from 'ids-enterprise-ng/lib';
+
+@Component({
+  templateUrl: 'contextual-action-panel-tabs-vertical.component.html'
+})
+
+export class ContextualActionPanelVerticalTabsComponent {
+  panel?: SohoContextualActionPanelRef<unknown>;
+
+  constructor() {
+  }
+
+  close() {
+    this.panel?.close();
+  }
+
+  onTabActivated(event: SohoTabsEvent) {
+    console.log(event.tab + ' TabsBasicDemoComponent.onTabActivated');
+  }
+}