+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@
 ### 10.2.0 Features
 
 - `[Cards]` Added Angular Wrapper Component for the IDS Cards. ([#1036](https://github.com/infor-design/enterprise-ng/issues/1036)) `EA`
+- `[ContextualActionPanel]` Added example of vertical tabs inside of Contextual Action Panel. ([#1065](https://github.com/infor-design/enterprise-ng/issues/1065)) `EA`
 - `[Lookup]` Added selected, afterpaging events and api setting to allow duplicate selected value to input element. ([#986](https://github.com/infor-design/enterprise-ng/issues/986))
 - `[Monthview]` Added Angular Wrapper Component for the IDS MonthView. ([#1051](https://github.com/infor-design/enterprise-ng/issues/1051)) `EA`